Update jkbms.py
The temperature fix for the JK BMS in v14.2 today was not working for me. Via Bluetooth I got -0°C as current temperature and with the code in version 14.2 I got 100°C.

For me

    self.to_temp(1, temp1 if temp1 <= 99 else (100 - temp1))
    self.to_temp(2, temp2 if temp2 <= 99 else (100 - temp2))
is working. Now I get 0°C in Venus OS. I have tested the temperature sensors now with and salt ice bath and my fingers. In the range from -16°C up to 30°C the measurement with the formula above is correct and corresponds the data in the JK-BMS android software. In the android software I could see -0°C and also 0°C and booth were translated to 0°C in Venus OS.